gnu: lightdm: Explicitely use a native 'bash' for 'wrap-program'.
Having to use a native bash for 'wrap-program' is fairly
unusual, but here it is the right thing, as it is for
the test suite only.  Placate the 'wrapper-inputs' linter.

Also, eventually we might consider letting 'wrap-program'
bail out if no bash for the host is found, even when
compiling natively.  But that's for later.

* gnu/packages/display-managers.scm
  (lightdm)[arguments]<#:phases>{pre-check}: Set #:sh
  argument of 'pre-check'

Signed-off-by: Maxim Cournoyer <maxim.cournoyer@gmail.com>