gnu: qtwebengine-5: Update to 5.15.10.
* gnu/packages/qt.scm (qtwebengine-5): Update to 5.15.10.
[source]: Update snippet to preserve the whole openh264 third party directory.

Series-to: 65280@debbugs.gnu.org
Cover-letter:
Refresh Qt 5 and 6 packages
This series refreshes our core Qt packages to their latest releases,
5.15.10 and 6.5.2.

It also adds some tooling for the Qt team in the form of two manifests to
select all the core Qt 5 packages or the other core packages, that builds on
top of the series posted at https://issues.guix.gnu.org/65230.

With this new tooling, it took less than two days to update our complete core
Qt packages collection, with the most time spent on qtwebengine (it has a
large allow-list of 3rd party libraries that needed to be tuned), as well as
investigating test failures in a few Qt 6.5.2 packages.
END
Series-version: 2