gnu: upower: Update to 1.90.2.
Upower has an extra hardware description file for use with udev.  It should be
installed under upower’s output, not eudev’s.

A test is disabled, for calling an unimplemented bluez method.  The
conditional 'adjust-test-for-excess-precision phase is included in the new
'adjust-test-suite, so that the list of phases is always the same.

* gnu/packages/gnome.scm (upower): Update to 1.90.2.
[#:configure-flags]: Set udevhwdbdir.
[#:phases] <adjust-test-for-excess-precision>: Add this phase
unconditionnally
<disable-failing-test>: Disable test_bluetooth_hidpp_mouse.
[inputs]: Remove dbus-glib.
[native-inputs]: Add dbus.